Split UndoManager into interface and JabRefUndoManager
Roughly 120 classes hold an undo handle, and almost all of them do one
thing with it: edit the library and hand the change over. Undoing,
redoing, asking whether the library differs from the last saved
position and subscribing to stack changes are the business of five
classes. Passing the whole manager to the rest handed every field
editor, cleanup and import task the ability to rewrite the user's
history, when all any of them does is describe what it just changed.

`UndoManager` is now the recording interface — `addEdit(BibChange)`,
`addEdit(String, Consumer)`, `applyEdit(BibChange)` — and
`JabRefUndoManager` is the implementation, following the shape JabRef
already uses for DialogService / JabRefDialogService. The class header
argued against a separate recording type on the grounds that it "would
mean threading a second handle everywhere the first one already goes";
that holds against a second object and not against an interface on the
same one, and the header now says so.

Keeping the interface under the old name is what makes this small. The
~118 recording clients keep the type name and the handle name they
already had, so the diff is 21 files rather than 133, and afterburner
keeps resolving injection sites because the registered key still
matches the declared type. `JabRefGUI` registers the one instance under
both keys, for the few classes that ask for the implementation.

Only five classes call anything beyond recording: GuiUndoManager,
UndoAction, RedoAction, EditAction and LibraryTab (SaveDatabaseAction
reaches markUnchanged through LibraryTab's getter). Sixteen further
classes hold the implementation and call nothing on it at all — they
carry it to reach one of those five, which is P12's problem, not an
interface gap: what they would need is undo/redo/canUndo/hasChanged,
which is the whole class again.
