+++ jabgui/src/main/java/org/jabref/gui/JabRefGUI.java
@@ -52,6 +52,7 @@ import org.jabref.logic.remote.RemotePreferences;
 import org.jabref.logic.remote.server.RemoteListenerServerManager;
 import org.jabref.logic.search.sqlbased.IndexManager;
 import org.jabref.logic.search.sqlbased.PostgresServer;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.logic.undo.UndoManager;
 import org.jabref.logic.util.BuildInfo;
 import org.jabref.logic.util.FallbackExceptionHandler;
@@ -90,7 +91,7 @@ public class JabRefGUI extends Application {
     private static FileUpdateMonitor fileUpdateMonitor;
     private static StateManager stateManager;
     private static ThemeManager themeManager;
-    private static UndoManager undoManager;
+    private static JabRefUndoManager undoManager;
     private static TaskExecutor taskExecutor;
     private static ClipBoardManager clipBoardManager;
     private static final String BIBTEX_EDITOR_FONT_RESOURCE = "fonts/JetBrainsMono-Regular.ttf";
@@ -238,8 +239,11 @@ public class JabRefGUI extends Application {
         );
         Injector.setModelOrService(ThemeManager.class, themeManager);
 
-        JabRefGUI.undoManager = new UndoManager();
+        JabRefGUI.undoManager = new JabRefUndoManager();
+        // Two keys, one instance: almost everything asks for the recording interface, while the
+        // few classes that build the undo UI ask for the implementation that can drive the stacks.
         Injector.setModelOrService(UndoManager.class, undoManager);
+        Injector.setModelOrService(JabRefUndoManager.class, undoManager);
 
         JabRefGUI.dialogService = new JabRefDialogService(mainStage);
         Injector.setModelOrService(DialogService.class, dialogService);
