+++ jabgui/src/main/java/org/jabref/gui/maintable/RightClickMenu.java
@@ -36,7 +36,7 @@ import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.preview.CitationStylePreviewLayout;
 import org.jabref.logic.shared.DatabaseLocation;
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.logic.util.TaskExecutor;
 import org.jabref.logic.util.io.FileUtil;
 import org.jabref.model.database.BibDatabaseContext;
@@ -51,7 +51,7 @@ public class RightClickMenu {
                                      DialogService dialogService,
                                      StateManager stateManager,
                                      GuiPreferences preferences,
-                                     UndoManager undoManager,
+                                     JabRefUndoManager undoManager,
                                      ClipBoardManager clipBoardManager,
                                      TaskExecutor taskExecutor,
                                      JournalAbbreviationRepository abbreviationRepository,
