+++ jabgui/src/main/java/org/jabref/gui/sidepane/SidePane.java
@@ -16,7 +16,7 @@ import org.jabref.gui.clipboard.ClipBoardManager;
 import org.jabref.gui.preferences.GuiPreferences;
 import org.jabref.logic.ai.AiService;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.logic.util.TaskExecutor;
 import org.jabref.model.entry.BibEntryTypesManager;
 import org.jabref.model.util.FileUpdateMonitor;
@@ -40,7 +40,7 @@ public class SidePane extends VBox {
                     FileUpdateMonitor fileUpdateMonitor,
                     BibEntryTypesManager entryTypesManager,
                     ClipBoardManager clipBoardManager,
-                    UndoManager undoManager) {
+                    JabRefUndoManager undoManager) {
         this.stateManager = stateManager;
         this.preferences = preferences;
         this.viewModel = new SidePaneViewModel(
