+++ jabgui/src/main/java/org/jabref/gui/undo/RedoAction.java
@@ -9,14 +9,14 @@ import org.jabref.gui.LibraryTab;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.actions.SimpleCommand;
 import org.jabref.logic.l10n.Localization;
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 
 import static org.jabref.gui.actions.ActionHelper.needsDatabase;
 
 public class RedoAction extends SimpleCommand {
     private final Supplier<LibraryTab> tabSupplier;
     private final DialogService dialogService;
-    private final UndoManager undoManager;
+    private final JabRefUndoManager undoManager;
 
     public RedoAction(Supplier<LibraryTab> tabSupplier, GuiUndoManager guiUndoManager, DialogService dialogService, StateManager stateManager) {
         this.tabSupplier = tabSupplier;
