+++ jabgui/src/main/java/org/jabref/gui/walkthrough/declarative/sideeffect/OpenLibrarySideEffect.java
@@ -13,7 +13,7 @@ import org.jabref.gui.preferences.GuiPreferences;
 import org.jabref.gui.walkthrough.Walkthrough;
 import org.jabref.logic.ai.AiService;
 import org.jabref.logic.importer.OpenDatabase;
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.logic.util.TaskExecutor;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.BibEntryTypesManager;
@@ -79,7 +79,7 @@ public class OpenLibrarySideEffect implements WalkthroughSideEffect {
                     stateManager,
                     Injector.instantiateModelOrService(FileUpdateMonitor.class),
                     Injector.instantiateModelOrService(BibEntryTypesManager.class),
-                    Injector.instantiateModelOrService(UndoManager.class),
+                    Injector.instantiateModelOrService(JabRefUndoManager.class),
                     Injector.instantiateModelOrService(ClipBoardManager.class),
                     Injector.instantiateModelOrService(TaskExecutor.class)
             );
