+++ jabgui/src/test/java/org/jabref/gui/edit/ManageKeywordsViewModelTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryPreferences;
 import org.jabref.model.entry.field.StandardField;
@@ -25,7 +25,7 @@ import static org.mockito.Mockito.when;
 class ManageKeywordsViewModelTest {
 
     private final BibEntryPreferences bibEntryPreferences = mock(BibEntryPreferences.class);
-    private final UndoManager undoManager = new UndoManager();
+    private final JabRefUndoManager undoManager = new JabRefUndoManager();
     private ManageKeywordsViewModel keywordsViewModel;
     private List<BibEntry> entries;
