+++ jabgui/src/test/java/org/jabref/gui/fieldeditors/optioneditors/LanguageEditorViewModelTest.java
@@ -9,7 +9,7 @@ import org.jabref.logic.FilePreferences;
 import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.logic.integrity.FieldCheckers;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
-import org.jabref.logic.undo.UndoManager;
+import org.jabref.logic.undo.JabRefUndoManager;
 import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.database.BibDatabaseMode;
 import org.jabref.model.entry.Langid;
@@ -39,7 +39,7 @@ public class LanguageEditorViewModelTest {
                 suggestionProvider,
                 BibDatabaseMode.BIBLATEX,
                 fieldCheckers,
-                new UndoManager()
+                new JabRefUndoManager()
         );
     }
